feat: 두 파일 이상에서 쓰이는 동일한 구조의 타입 common 폴더에서 관리
