+++ src/apis/common/type.d.ts
@@ -1,0 +1,10 @@
+export interface UserSummary {
+  id: number;
+  nickname: string;
+  email: string;
+}
+
+export interface DashboardSummary {
+  id: number;
+  title: string;
+}
